+++ Assets/Scripts/Gameplay/Configs/Attack/Attacks/PlayerDarkAttack.cs
@@ -13,6 +13,6 @@ namespace Gameplay.Configs.Attacks
         public VFXView ExplosionVFX;
         public float Radius = 5;
         public float TimeBeforeHit = 1;
-        public float Damage = 10;
+        public int Damage = 10;
     }
 }
